ADV-2024: Adds day 3
Adds Day 3. Using regex always feels a bit like cheating here
but hey-ho it made this a lot easier. TIL about `|` vs `||`
for regex expressions. This made it a lot easier as regex is
evaluated from the start of the string to the end, so no need
to keep track of the position in part 2. The if bit are a bit
gross but they worked!

Rating: ⭐️ + ½ - I learned a new regex thing