ADV-2024: Adds day 4
This one was a bit tricky but was a rare case of part 1
being trickier than part 2. I think what made part 2
a lot easier was the ability to exploit symmetry around
the A char. The only real snag was spotting the case of
the same letters opposite each other.

Part 1 felt tricker mainly because what I misread
the puzzle to assume tracking xmas values allowed
any direction, so was getting a massively too high
value. I wrote a lot more code for this part than
part 2.

Rating: ⭐️⭐️⭐️