ADV-2024: Adds Day 5
Writing a custom comparator made this really easy,
Java for the win I guess. Really the easy part
was the dead simple comparison logic which just
involved looking up the strings in rule set.

This approach really held up in part 2 too, so
was needed no logic changes to do to. Overall
quite simple, but was luck with my approach

Rating: ⭐️⭐️